using form entered value instead of raw data element value
